chore: update the version of `screen_layer`

+++ kernel/src/graphics/screen/layer.rs
@@ -34,13 +34,13 @@ where
     get_controller().lock().edit_layer(id, f)
 }
 
-pub fn slide(id: screen_layer::Id, new_top_left: Vec2<isize>) -> Result<(), screen_layer::Error> {
+pub fn slide(id: screen_layer::Id, new_top_left: Vec2<i32>) -> Result<(), screen_layer::Error> {
     get_controller().lock().slide_layer(id, new_top_left)
 }
 
 pub fn set_pixel(
     id: screen_layer::Id,
-    coord: Vec2<isize>,
+    coord: Vec2<u32>,
     color: Option<RGB8>,
 ) -> Result<(), screen_layer::Error> {
     get_controller().lock().set_pixel(id, coord, color)
